+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Also check this project's [releases](https://github.com/powerhome/redis-operator
 
 ## Unreleased
 
-- [Allow different ports on BootstrapNode and RF]()
+- [Allow different ports on BootstrapNode and RF](https://github.com/powerhome/redis-operator/pull/64)
 
 ## [v4.1.0]  - 2025-01-16